+++ src/tooling/piral-cli/src/commands.ts
@@ -419,6 +419,8 @@ const allCommands: Array<ToolCommand<any>> = [
         .default('bundler', availableBundlers[0])
         .string('app')
         .describe('app', 'Sets the name of the Piral instance.')
+        .string('app-dir')
+        .describe('app-dir', 'Sets the path to a custom Piral instance for serving.')
         .string('base')
         .default('base', process.cwd())
         .describe('base', 'Sets the base directory. By default the current directory is used.')
@@ -432,6 +434,7 @@ const allCommands: Array<ToolCommand<any>> = [
         hmr: args.hmr as boolean,
         bundlerName: args.bundler as string,
         optimizeModules: args['optimize-modules'] as boolean,
+        appInstanceDir: args['app-dir'] as string,
         app: args.app as string,
         logLevel: args['log-level'] as LogLevels,
         open: args.open as boolean,